My, it's working!
I wouldn't dare to call it alpha, but it do something. When you start,
you're on 4th (Auctionator) tab, and here you can post an item and make
an auction with it; Auctionator will count a price appropriately. You
still can't switch on 4th tab manually, and it's performance still bad.